add current user check

+++ HomewOurK.WebAPI/Controllers/UsersController.cs
@@ -1,5 +1,6 @@
 ﻿using HomewOurK.Application.Interfaces;
 using HomewOurK.Domain.Entities;
+using HomewOurK.WebAPI.Helpers;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authorization;
@@ -11,9 +12,16 @@ namespace HomewOurK.WebAPI.Controllers
 	[Route("api/[controller]")]
 	[ApiController]
 	[Authorize]
-	public class UsersController(IUserService userService) : ControllerBase
+	public class UsersController : ControllerBase
 	{
-		private readonly IUserService _userService = userService;
+		private readonly IUserService _userService;
+		private readonly IHttpContextAccessor _httpContextAccessor;
+
+		public UsersController(IUserService userService, IHttpContextAccessor httpContextAccessor)
+		{
+			_userService = userService;
+			_httpContextAccessor = httpContextAccessor;
+		}
 
 		[HttpGet]
 		public IActionResult GetAll()
@@ -28,21 +36,35 @@ namespace HomewOurK.WebAPI.Controllers
 		[HttpGet("GetUsers")]
 		public IActionResult GetUsers(int groupId)
 		{
-			var users = _userService.GetUsersByGroupId(groupId);
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
 
-			if (users.Any())
-				return Ok(users);
-			return NotFound("No users were found for the group with id = " + groupId);
+			if (_userService.UserInGroup(groupId, email))
+			{
+				var users = _userService.GetUsersByGroupId(groupId);
+
+				if (users.Any())
+				{
+					foreach (var user in users)
+						user.Password = null;
+					return Ok(users);
+				}
+				return NotFound("No users were found for the group with id = " + groupId);
+			}
+			return BadRequest();
 		}
 
 		[HttpGet("GetUser")]
-		public IActionResult GetUser(int userId)
+		public IActionResult GetUser()
 		{
-			var users = _userService.GetUserById(userId);
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+			var user = _userService.GetUserByEmail(email);
 
-			if (users != null)
-				return Ok(users);
-			return NotFound($"The user with id = {userId} wasn't found");
+			if (user != null)
+			{
+				user.Password = null;
+				return Ok(user);
+			}
+			return NotFound($"The user with email = {email} wasn't found");
 		}
 
 		[HttpPost("Register")]
@@ -51,7 +73,10 @@ namespace HomewOurK.WebAPI.Controllers
 		{
 			user.Id = 0;
 			if (_userService.AddUser(user))
+			{
+				user.Password = null;
 				return Ok(user);
+			}
 			return BadRequest("A user with this email already exists");
 		}
 
@@ -83,22 +108,38 @@ namespace HomewOurK.WebAPI.Controllers
 		[HttpPatch]
 		public IActionResult UpdateUser(User user)
 		{
-			if (_userService.UpdateUser(user))
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+			var cookieUser = _userService.GetUserByEmail(email);
+
+			if (cookieUser != null && cookieUser.Email != user.Email)
+				return Unauthorized();
+
+			if (user != null && cookieUser != null)
 			{
-				user.Password = null;
-				return Ok(user);
+				user.Id = cookieUser.Id;
+				user.Email = cookieUser.Email;
+				user.Password = cookieUser.Password;
+				if (_userService.UpdateUser(user))
+				{
+					user.Password = null;
+					return Ok(user);
+				}
 			}
 			return BadRequest();
 		}
 
 		[HttpDelete]
-		public IActionResult DeleteUser(User user)
+		public IActionResult DeleteUser()
 		{
-			if (_userService.DeleteUser(user))
-			{
-				user.Password = null;
-				return Ok(user);
-			}
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+			var user = _userService.GetUserByEmail(email);
+
+			if (user != null)
+				if (_userService.DeleteUser(user))
+				{
+					user.Password = null;
+					return Ok(user);
+				}
 			return BadRequest();
 		}
 	}
